Remove content_moderation from defaults. (#267)
* Remove content_moderation from defaults.

Workflow isn't configured within services so an empty config doesn't
hepl (and breaks workspaces as such). It should be added with
localgov_workflows.

* Fix previously broken tests by changing theme to one that has class on tabs.

The JS drag tests were failing as the tab with the field to drag to was
not being selected. The class being looked for is supplied by claro.

* Add basic Workflow integration test to ensure form fields.

Rest of integration test, access etc. included in LocalGov Workflows.

* Coding standards fixes.